Transceiver52M: Add FCCH based frequency correction
Enable frequency detection and correction by buffering the previous
frame to allow FCCH measurement and compensation after frame timing is
locked using the SCH. When the SCH is detected and symbol timing
matched, measure the FCCH burst from one frame prior and compensate by
baseband tuning the DDC on the device. Avoid appying frequency
corrections to the RF portion due to possible tuning delays, which is
not an issue with DDC tuning.

Signed-off-by: Thomas Tsou <tom@tsou.cc>