Add Device Definition for the OC Connect1 SDR
This patch forward ports the relevant parts of a patch
from OpenCellular for osmo-trx versions previous to commit
1fb0ce67d86067d81bd0f8d18b8c11890e36e755
so that osmo-trx-uhd recognizes the hardware, somewhat documented here:

https://github.com/Telecominfraproject/OpenCellular/blob/master/electronics/radio/SDR/

Note: I'm not very familiar with the hardware. It requires a patch to the
Ettus UHD driver to load the correct FPGA fw in which one call to
check_fpga_compat() is commented.
Otherwise mostly changes identifiers and log messages to "OCR01".

The usb device is reported as Vid:2500 Pid:0020 (Ettus,B200) but
I'm not sure how compatible it is. The fpga FW is different.