contrib/jenkins: remove broken chroot + qemu code
The mirror where it would download images from is not online anymore,
and it looks like this whole block has not been executed in years (we
always test the arm builds on native arm machines nowadays, and run all
builds in docker).

Replace the whole block with a simple check that ensures when arm flags
are used, that the build is actually running on an arm machine.

Related: OS#6549
Change-Id: I2e204e4a0a7dfcf32f011ed6fb403b564f8155a7